requirements.txt
streamlit
pillow
tensorflow
transformers
torch
rembg
opencv-python
numpy
python-dotenv
tqdm
matplotlib
albumentations
openai